requirements.txt
boto3
botocore
awswrangler
awswrangler[redshift]
pandas
openpyxl
pyspark
mlflow
apache-airflow
gradio
flask
psutil
tqdm